LeanDog - Intelligent Engineering with AI/Project/trello-extract/src/clients/trello_client.py
from trello import TrelloClient

from src.settings import Settings


def get_trello_client(settings: Settings) -> TrelloClient:
    return TrelloClient(
        api_key=settings.trello_api_key,
        api_secret=settings.trello_api_token,
    )
